requirements.txt
aiofiles==23.2.1
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
asgiref==3.8.1
asttokens==2.4.1
asyncer==0.0.2
attrs==23.2.0
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.3
bidict==0.23.1
bs4==0.0.2
build==1.2.1
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
chainlit==1.0.502
chardet==5.2.0
charset-normalizer==3.3.2
chevron==0.14.0
chroma-hnswlib==0.7.3
chromadb==0.4.24
click==8.1.7
cmake==3.29.0.1
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.2.1
cryptography==42.0.5
cycler==0.12.1
dataclasses-json==0.5.14
dataclasses-json-speakeasy==0.5.11
debugpy==1.8.1
decorator==5.1.1
Deprecated==1.2.14
distro==1.9.0
effdet==0.4.1
emoji==2.11.0
et-xmlfile==1.1.0
executing==2.0.1
faiss-cpu==1.8.0
fastapi==0.110.1
fastapi-socketio==0.0.10
ffmpy==0.3.2
filelock==3.13.3
filetype==1.2.0
flatbuffers==24.3.25
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.3.1
google-auth==2.29.0
googleapis-common-protos==1.63.0
gpt4all==2.4.0
gradio==4.26.0
gradio_client==0.15.1
greenlet==3.0.3
grpcio==1.62.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.22.2
humanfriendly==10.0
idna==3.6
importlib-metadata==7.0.0
importlib_resources==6.4.0
iopath==0.1.10
langchain==0.1.14
langchain-community==0.0.32
langchain-core==0.1.41
langchain-text-splitters==0.0.1
langdetect==1.0.9
langsmith==0.1.40
layoutparser==0.3.4
Lazify==0.4.0
literalai==0.0.500
lxml==5.2.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
matplotlib==3.8.4
matplotlib-inline==0.1.6
mdurl==0.1.2
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
nltk==3.8.1
numpy==1.26.4
oauthlib==3.2.2
olefile==0.47
ollama==0.1.8
omegaconf==2.3.0
onnx==1.16.0
onnxruntime==1.15.1
openai==1.16.2
opencv-python==4.9.0.80
openpyxl==3.1.2
opentelemetry-api==1.24.0
opentelemetry-exporter-otlp==1.24.0
opentelemetry-exporter-otlp-proto-common==1.24.0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
opentelemetry-exporter-otlp-proto-http==1.24.0
opentelemetry-instrumentation==0.45b0
opentelemetry-instrumentation-asgi==0.45b0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-proto==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-semantic-conventions==0.45b0
opentelemetry-util-http==0.45b0
orjson==3.10.0
outcome==1.3.0.post0
overrides==7.7.0
packaging==23.2
pandas==2.2.1
parso==0.8.4
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.0
pikepdf==8.15.0
pillow==10.3.0
pillow_heif==0.16.0
pip==24.0
platformdirs==4.2.0
portalocker==2.8.2
posthog==3.5.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
pulsar-client==3.4.0
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycocotools==2.0.7
pycparser==2.22
pydantic==2.6.4
pydantic_core==2.16.3
pydub==0.25.1
Pygments==2.17.2
PyJWT==2.8.0
pypandoc==1.13
pyparsing==3.1.2
pypdf==4.2.0
pypdfium2==4.28.0
PyPika==0.48.9
pyproject_hooks==1.0.0
pyreadline3==3.4.1
PySocks==1.7.1
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python-docx==1.1.0
python-dotenv==1.0.1
python-engineio==4.9.0
python-graphql-client==0.4.3
python-iso639==2024.2.7
python-magic==0.4.27
python-multipart==0.0.9
python-pptx==0.6.23
python-socketio==5.11.2
pytz==2024.1
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.2
rapidfuzz==3.8.0
referencing==0.34.0
regex==2023.12.25
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.7.1
rpds-py==0.18.0
rsa==4.9
ruff==0.3.5
safetensors==0.4.2
scipy==1.13.0
selenium==4.19.0
semantic-version==2.10.0
setuptools==65.5.0
shellingham==1.5.4
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.29
stack-data==0.6.3
starlette==0.37.2
sympy==1.12
syncer==2.0.3
tabulate==0.9.0
tenacity==8.2.3
tiktoken==0.6.0
timm==0.9.16
tokenizers==0.15.2
tomli==2.0.1
tomlkit==0.12.0
toolz==0.12.1
torch==2.2.2
torchvision==0.17.2
tornado==6.4
tqdm==4.66.2
traitlets==5.14.2
transformers==4.39.3
trio==0.25.0
trio-websocket==0.11.1
typer==0.12.1
typing_extensions==4.11.0
typing-inspect==0.9.0
tzdata==2024.1
unstructured==0.13.2
unstructured-client==0.18.0
unstructured-inference==0.7.25
unstructured.pytesseract==0.3.12
uptrace==1.24.0
urllib3==2.2.1
uvicorn==0.25.0
watchfiles==0.20.0
wcwidth==0.2.13
websocket-client==1.7.0
websockets==11.0.3
wrapt==1.16.0
wsproto==1.2.0
xlrd==2.0.1
XlsxWriter==3.2.0
yarl==1.9.4
zipp==3.18.1